project/Client.cpp: moved to .h and .cpp
