+++ project/compile.sh
@@ -1,2 +1,2 @@
 clang++ -Iboost_1_85_0 -Iinclude -Lboost_1_85_0/stage/lib -lsodium -std=c++23 src/ExchangeCore.cpp src/MatchingEngine.cpp src/EncryptionUtils.cpp src/UserUtils.cpp Exchange.cpp -o exchange
-clang++ -Iboost_1_85_0 -Lboost_1_85_0/stage/lib -std=c++23 Client.cpp -o client
+clang++ -Iboost_1_85_0 -Iinclude -Lboost_1_85_0/stage/lib -std=c++23 src/Client.cpp Launcher.cpp -o client
